move to metta.common.util.module
