+++ metta/cogworks/curriculum/task_generator.py
@@ -16,8 +16,8 @@ from pydantic import (
 from typing_extensions import Generic
 
 from metta.common.config import Config
+from metta.common.util.module import load_symbol
 from metta.mettagrid.mettagrid_config import EnvConfig
-from metta.utils.module import load_symbol
 
 if TYPE_CHECKING:
     from metta.cogworks.curriculum.curriculum import CurriculumConfig
